[css-text] A leading white-space should break before handling overflow
Leading white-spaces are indeed breaking opportunities that should
prevent, if there are no other css properties forcing it, breaking text
in the middle of a word when honoring the word-wrap/overflow-wrap CSS
property.

We are doing so if the leading white-space sequence is longer than 1
character, but when we have a single leading white-space, we are missing
that breaking opportunity and we may lead to cases, like the one
described in the bug.

However, this breaking opportunity should be considered together
with other provided by the word-break CSS property (eg, break-word or
break-all).

There is a discussion [1] in the CSS WG github about this issue, which
provides an interoperability analysis of a few cases where a single
leading white-space character may produce different results, depending
on the browser and the combination of line-breaking properties.

Bug: 854624
Change-Id: I49ec8282e899a32990662c104b48cab04057b909